requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
mysql-connector-python
sqlalchemy
asyncmy
pinecone
langgraph
langchain
langchain-google-genai
google-generativeai
google-genai
beautifulsoup4
python-dotenv
requests
aiohttp
feedparser